// client/src/State/actionTypes.ts
enum authConstant {
  USER_SIGNUP_REQUEST = "USER_SIGNUP_REQUEST",
  USER_SIGNUP_SUCCESS = "USER_SIGNUP_SUCCESS",
  USER_SIGNUP_FAILURE = "USER_SIGNUP_FAILURE",
  USER_LOGIN_REQUEST = "USER_LOGIN_REQUEST",
  USER_LOGIN_SUCCESS = "USER_LOGIN_SUCCESS",
  USER_LOGIN_FAILURE = "USER_LOGIN_FAILURE",
}

enum todoConstant {
  CREATE_TODO_REQUEST = "CREATE_TODO_REQUEST",
  CREATE_TODO_SUCCESS = "CREATE_TODO_SUCCESS",
  CREATE_TODO_FAILURE = "CREATE_TODO_FAILURE",
  UPDATE_TODO_REQUEST = "UPDATE_TODO_REQUEST",
  UPDATE_TODO_SUCCESS = "UPDATE_TODO_SUCCESS",
  UPDATE_TODO_FAILURE = "UPDATE_TODO_FAILURE",
  DELETE_TODO_REQUEST = "DELETE_TODO_REQUEST",
  DELETE_TODO_SUCCESS = "DELETE_TODO_SUCCESS",
  DELETE_TODO_FAILURE = "DELETE_TODO_FAILURE",
  GET_TODO_REQUEST = "GET_TODO_REQUEST",
  GET_TODO_SUCCESS = "GET_TODO_SUCCESS",
  GET_TODO_FAILURE = "GET_TODO_FAILURE",
}

export { authConstant, todoConstant };